# lxdweb/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Session 的认证",
        "title": "LXD Web 管理平台",
        "termsOfService": "https://github.com/xkatld/zjmf-lxd-server",
        "contact": {
            "name": "xkatld",
            "url": "https://github.com/xkatld/zjmf-lxd-server",
            "email": "xkatld@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/xkatld/zjmf-lxd-server/blob/main/LICENSE"
        },
        "version": "1.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/ipv6": {
            "get": {
                "description": "查询所有IPv6绑定信息，支持按节点过滤",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "获取IPv6绑定列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回IPv6绑定列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "为指定容器创建新的IPv6地址绑定",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "创建IPv6绑定",
                "parameters": [
                    {
                        "description": "IPv6绑定参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateIPv6Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6-sync/all": {
            "post": {
                "description": "启动所有活跃节点的IPv6绑定同步任务",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "同步所有节点的IPv6绑定",
                "responses": {
                    "200": {
                        "description": "同步任务已启动",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6-sync/status": {
            "get": {
                "description": "查询指定节点或所有节点的IPv6绑定同步状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "获取IPv6同步状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回同步状态",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6/cache": {
            "get": {
                "description": "直接从本地缓存数据库读取IPv6绑定信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "从缓存获取IPv6绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "容器名称",
                        "name": "hostname",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回缓存数据",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6/sync": {
            "post": {
                "description": "从lxdapi同步指定节点的IPv6绑定信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "同步IPv6绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "同步任务已启动",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "缺少参数",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6/tasks": {
            "get": {
                "description": "查询最近50条IPv6同步任务记录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "获取IPv6同步任务列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "节点ID",
                        "name": "node_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功返回任务列表",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "查询失败",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/ipv6/{id}": {
            "delete": {
                "description": "删除指定的IPv6地址绑定并清理数据库",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPv6管理"
                ],
                "summary": "删除IPv6绑定",
                "parameters": [
                    {
                        "type": "string",
                        "description": "绑定ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "绑定或节点不存在",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.CreateIPv6Request": {
            "type": "object",
            "required": [
                "container_hostname",
                "node_id"
            ],
            "properties": {
                "container_hostname": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "node_id": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}